moved parent expanded mobile components into single component page
